Use the newly discovered intMap! :D
