+++ samples/collision2d-roguelike/build.gradle.kts
@@ -3,7 +3,7 @@ import com.soywiz.korge.gradle.*
 apply<KorgeGradlePlugin>()
 
 korge {
-	id = "com.soywiz.samples.collision2d"
+	id = "com.soywiz.samples.collision2d.roguelike"
 
 	targetDefault()
 }
