+++ samples/collision2d-roguelike/src/commonMain/resources/32x32_ortho_dungeon_tile_Denzi111112-2.png
