+++ samples/collision2d-roguelike/src/commonMain/resources/CREDITS
