+++ samples/collision2d-roguelike/src/commonMain/resources/tilemap.tmx
